refactor styled comps,rename,  fix small issues
