+++ pages/index.js
@@ -1,5 +1,5 @@
 import ReviewList from '../src/components/ReviewList/ReviewList';
-import { ListGrid } from '../src/components/ReviewList/ListGrid.styled';
+import { ListGrid } from '../src/UI/ListGrid.styled';
 import Form from '../src/components/Form/Form';
 import Modal from '../src/components/Form/Modal';
