+++ src/UI/Legend.styled.js
